Reduce the font size when printing.
What this actually does is increase the maximum text width when
printing. The browser takes the page and scales it to fit horizontally
to the page so when we have a max width of 65ch the browser will scale
up our UI so that the elements fit the width of the page. By increasing
the maximunm width we reduce the amount the browser needs to scale up to
fit onto an 8.5" wide page.