Renames SectionComponent -> SimpleSectionComponent.
SectionComponent sounds like it should be a common base class for all
other components. GetStartedComponent, OverviewComponent,
StepByStepComponent, etc. are all sections of a resource. The intention
here is to make it more clear that the <sbdl-simple-section> is not a
base for other sections but just a common implementation for sections
with simple content.