Include resource type in the name of  resource-type-specific subcomponents.
All resource-related components currently live in the same
ResourceModule. This leads to some naming conflicts. For example both
Professional Learning, Strategy, and Connections Playlist resources have
"overview" sections and these all differ enough from each other that it
makes more sense for each resource type to have a separate overview
component rather than a single shared component for all three. However,
the Professional Learning-specific overview implementation already took
the names `OverviewComponent` and `sbdl-overview`. So when implementing
Strategies we created `StrategyOverviewComponent`. At a glance this is
misleading because it implies that `OverviewComponent` is more generally
applicable than it actually is. We had similar issues with the naming of
other sections in the Instructional and Professional Learning resources.

Two solutions are immediately apparent:

1. Create resource-specific submodules which only expose the main resource
   and content components. In this approach we could have multiple
   `OverviewComponents` because each would be encapsulated in their
   specific resource module.

2. Adopt a new convention where we include the name of the parent
   resource type in the name of all resource-specific sub-components.

I decided to go with the second approach for three reasons:

1. creating resource-specific submodules would require a fair amount of
   restructuring of the folder structure;

2. creating resource-specific submodules would require a significant
   refactor of ResourceModule into ResourceCommonModule removing all
   the resource-specific components;

3. while the symmetry of each resource type getting it's own
   `OverviewComponent` is attractive I think it would actually make the
   content components less clear. It would mean we have multiple
   possible answers to questions like "which component does `sbdl-overview`
   represent?" Answering this question would require knowing which
   module the current component is in and that requires inspecting each
   of the module definitions to answer conclusively.

   By contrast if you are working on the `StrategyContentComponent`
   template and see a reference to `sbdl-professional-overview` it is
   immediately apparent that the wrong component is being referenced.

For consistency all sub-components have been renamed to follow this new
naming pattern. It should still be the case the all components follow
the rule that a component with class name `FooBarComponent` has the
selector name of `sbdl-foo-bar-component` and lives in files named
`foo-bar.component.{ts,html,scss}`.