Add tabindex support for sbdl-button-icon elements.
There are cases when we need to remove the button element itself from
the tab order. For example when we have a link inside the button. In
this case we want the link because we want all of the browser's native
behavior for links (vs. a "psuedo-link" implemented with a button) and
that is what it is semantically (again taking advantage of default
accessibility behaviors). And we want to use the button component to get
the visual cues and pleasant ripple effect. However this nows gives us
two tab targets: the button and the link. furthermore, the button itself
does not have click behavior defined, so if a user tabbing through
elements lands on the button and hits <Enter> nothing will happen.

The solution we're currently using is to remove the button from the tab
order with `tabindex="-1"` (hence this change). This leaves the user
with only one tab target for this element: the link. Clicking or hitting
<Enter> behaves as links normally do. The only drawback with this
approach is that the link receiving focus does not currently trigger the
background highlight to appear, whereas hovering over the link (and
button) still does.