DL-186 Fix padding for lists in Read More components.
The ReadMoreComponent is intended to add a "more/less" link when the
wrapped text contains more than two lines. The way we implement this is
based on interogating the element for it's total height. If the total
height is greater than two lines should be based on the known font size
then we add the more/less link. This whole calculation depends on there
not being excess padding internally within the wrapped text.

Additionally, we now hide the more/less link from screen readers (as
they will just read the whole block).