DL-198 Fixes metadata formating in printed layouts.
Previously the resource header (with the title, etc) was part of the
same element as the main resource content. Positioning the metadata
after the header but before the main content required manually adding
padding and absoutely positioning the metadata block. To do this
properly we would need to inspect the actiual sizes of these elements at
print-time to accurately calculate the padding needed.

This commit bypasses all of that by instead breaking the resource header
out as a separate component and switching to a grid layout so that we
can let the browser handle all of the layout calculations.