DL-229 Best-effort fix for footer formating in printed resources.
The actual fix for this requires technology that's not available yet.
CSS Paged Media provides the means to format footers in the way that we
wish but is not yet implemented in our target browsers. For the time
being this adjusts some of the margins and positioning of the footer to
better situate it on the page but there is norealy way to create a
persistent margin for the page content that does not also serve as a
margin for the footer.

In other words, if I add a 2cm margin to the bottom of the page the
footer element must also respect that margin. Trying to put the footer
within the margin pushes it onto the next page. Adding margin on the
footer itself doesn't work because the footer is fixed (has to be to
appear on every page) and fixed elements are not considered when laying
out the primary page flow (so the footer margin doesn't affect the body
content). Adding a margin at the bottom of the content only affects the
last page. Using `@page { margin: 2cm; }` causes the issue described in
the first paragraph (footer is affected).