Move the API URL prefix out of config and into calling code.
When search results are returned by the API, the API provides links to
the first, last, and next pages of results. These links are formatted as
relative paths from the root of the domain. If they are passed into our
DataService it previously would result in urls like:

    https://dev.dl.webapp.sb.org/api/api/search...

With that URL prefix of `api` being repeated. To avoid this we're
changing our internal config to match what the API is using.