Update resource teaser model mapping to reflect current API model.
Previously the teaser resources had different names for the `overview`
section of the resource. This differed from full resources, which all
used `overview` as the field name. This discrepancy has been corrected
in the API and now all the teaser resources use `overview`. This change
removes the prior custom mappings for teaser resources.